requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.0.9
django-appconf==1.0.3
django-cas-client==1.5.0
django-compressor==2.2
django-constance==2.4.0
django-extensions==2.1.6
django-ical==1.5
django-loginas==0.3.6
django-phonenumber-field==2.3.1
django-picklefield==2.0
django-post-office==3.1.0
django-sass-processor==0.7.2
django-webpush==0.2.7
http-ece==1.1.0
icalendar==4.0.3
idna==2.8
jsonfield==2.0.2
libsass==0.18.0
phonenumbers==8.10.9
psycopg2-binary==2.8.2
py-vapid==1.5.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
pywebpush==1.6.0
rcssmin==1.0.6
requests==2.21.0
rjsmin==1.0.12
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
whitenoise==4.1.2
